CI: Fix deploy-wasm trigger condition
The potential performance effect files are not tracked when deploying
Wasm, so they are now added to the file detection list. Additionally,
the repository_workflow event is now being listened to since most ELF
files are built and stored in the rv32emu-prebuilt repository. When a
release is published in rv32emu-prebuilt, it triggers the
repository_workflow event in rv32emu to deploy Wasm, ensuring that new
ELF or Linux images are properly synced.