Enable to boot on virtio block device
Previously, when booting on virtio block device
e.g., /dev/vda, stuck at the init process.

The root cause is virtio block interrupt cannot be handled
inside the trap handler stepping loop. Thus, nested interrupt
should be supported. This is a simple workaround to handle the
virtio block interrupt inside the stepping loop.
!!! USE MORE ELEGANT WAY TO HANDLE IT !!!

The PLIC's IRQ line is up to 32, 1 is for serial, so 30 more free
IRQ line can be used by virtio blocks. Above 30, the behaviour is
unknown. Thus, the name of the last virtio block is /dev/vdad.

The boot.sh is used to quick testing to adjust the number of
virtio blocks.

After boot the guestOS, Run mount command to check the rootfs mount type
- using initramfs, /dev/root on / type rootfs
- using virtio block device, should show /dev/root on / type ext4

Bug:
1. When running quake, doom, smolnes, the guest OS hangs and terminated
   by the hostOS.